Rewrite continues to continue.
Currently, I am keeping a similar system to what I had before, with multiple fragments and all of that. I'm just remaking them to be more specific to what they represent.

Effectively, instead of FragmentOne.java, we're going to have AnnouncementsFragment.java (or whatever it is that we end up wanting).

Some of the XML used for Announcement may be able to be reused later, such as card_item. Additionally, MyRecyclerViewAdapater.java can be reused too.