Fixed NullPointerException
I was attempting to populate the data ArrayList in AnnouncementsFragments before actually instantiating it, so I fixed that.

Currently, the Announcements tab has the appropriate number of cards containing text. The issue is that the Wildcat icon is underneath the text and a majority of the text is cut off.