Fixed Duplication Bug by calling data.clear() at start of populateData()
populateData() is called every time onCreateView() is called by an AnnouncementFragment. This happens fairly often. Effectively, with the way RecyclerView works and all, it happens a lot. That means that every single time populateData is called, all of this the data below is re-added to the data ArrayList. If I neglect to clear the ArrayList (I ensure that it isn't null to avoid a NullPointerException), there will be duplicated data in the ArrayList. Android obviously doesn't know any better than to create extra Cards out of this duplicated data, resulting in lots and lots of cards with the exact same data. Clearing the ArrayList each time the populateData() method is called ensures that there aren't any duplicates. Whether or not there's a better way to do this is beyond me at the moment, but this works currently and I'm fine with that.

The above block of texts is also in a comment above the data.clear() call.